Added support for reversed Edges.
Dot does directional layout based on the directionality of the edges
(i.e. top down).  This means that all arrows point downwards.  Sometime
in a dot layout you want arrows to point up.  To do this a reversed edge
swaps the order of the nodes connected by the edge and puts the arrow on
the tail rather than the head.